Natura Restful API:
RESTful API ~~ Srping-boot + Spring security + Oauth2:
~~DONE~~
Conection to a database with hibernate.
Authentication with credentials of a database: Authentication provider and UserDetailService.
Logout in memory with tokens via headers.
CORS and CRSF implemented.
TokenEnhancer addded to add information on tokens
~~NOT YET/ UNCOMPLETE~~
Bcryp implemented in Registration ~ not yet in login.

NaturaWEB on AngularJS:
~~DONE~~
Login and Register page with CSS.
Home with redirect in case of unlogged user.
Cache list of Products, list of Users and User detail.
~~NOT YET/UNCOMPLETE~~
Home without CSS.
CRUD of Products and Users